feat: Adding Schedule Notification to DB

+++ app/Http/Controllers/BookingController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use App\Models\Booking;
 use App\Http\Requests\StoreBookingRequest;
 use App\Http\Requests\UpdateBookingRequest;
+use App\Models\ScheduledNotification;
+use App\Notifications\BookingReminder1H;
+use Carbon\CarbonImmutable;
 use Illuminate\Http\RedirectResponse;
 
 class BookingController extends Controller
@@ -35,13 +38,32 @@ class BookingController extends Controller
     public function store(StoreBookingRequest $request): RedirectResponse
     {
         //$request->user()->bookings()->create($request->validated());
-        $request->user()->bookings()->create([
-           // 'start' => $request->validated('start'),
-            'start' => fromUserDateTime($request->validated('start')),
-            //'end' => $request->validated('end'),
-            'end' => fromUserDateTime($request->validated('end')),
+        //$request->user()->bookings()->create([
+        //    'start' => fromUserDateTime($request->validated('start')),
+        //    'end' => fromUserDateTime($request->validated('end')),
+        //]);
+
+        $booking = $request->user()->bookings()->create([
+            'start' => fromUserDateTime($request->validated('start'), $request->user()),
+            'end' => fromUserDateTime($request->validated('end'), $request->user()),
         ]);
 
+        $startTime = CarbonImmutable::parse(toUserDateTime($booking->start, $booking->user), $booking->user->timezone);
+
+        // Schedule 1H reminder
+        $oneHourTime = fromUserDateTime($startTime->subHour(), $booking->user);
+        if (now('UTC')->lessThan($oneHourTime)) {
+            $booking->user->scheduledNotifications()->create([
+                'notification_class' => BookingReminder1H::class,
+                'notifiable_id' => $booking->id,
+                'notifiable_type' => Booking::class,
+                'sent' => false,
+                'processing' => false,
+                'scheduled_at' => $oneHourTime,
+                'sent_at' => null,
+                'tries' => 0,
+            ]);
+        }
 
         return redirect()->route('booking.index');
     }
@@ -76,6 +98,38 @@ class BookingController extends Controller
             'end' => fromUserDateTime($request->validated('end')),
         ]);
 
+        $startTime = CarbonImmutable::parse(toUserDateTime($booking->start, $booking->user), $booking->user->timezone);
+
+        $hasScheduledNotifications = ScheduledNotification::query()
+            ->where('notifiable_id', $booking->id)
+            ->where('notifiable_type', Booking::class)
+            ->where('user_id', $booking->user_id)
+            ->exists();
+
+        // First we need to check if there are any already scheduled notifications
+        if ($hasScheduledNotifications) {
+            // Then in this example, we simply delete them. You can however update them if you want.
+            $booking->scheduledNotifications()
+                ->where('user_id', $booking->user_id)
+                ->delete();
+        }
+
+        // Since we are clearing the scheduled notifications, we need to create them again for the new date
+        // Schedule 1H reminder
+        $oneHourTime = fromUserDateTime($startTime->subHour(), $booking->user);
+        if (now('UTC')->lessThan($oneHourTime)) {
+            $booking->user->scheduledNotifications()->create([
+                'notification_class' => BookingReminder1H::class,
+                'notifiable_id' => $booking->id,
+                'notifiable_type' => Booking::class,
+                'sent' => false,
+                'processing' => false,
+                'scheduled_at' => $oneHourTime,
+                'sent_at' => null,
+                'tries' => 0,
+            ]);
+        }
+
         return redirect()->route('booking.index');
     }
 
@@ -87,6 +141,10 @@ class BookingController extends Controller
         abort_unless($booking->user_id === $request->user()->id, 404);
 
         $booking->delete();
+
+        $booking->scheduledNotifications()
+            ->where('user_id', $booking->user_id)
+            ->delete();
 
         return redirect()->route('booking.index');
     }
